Updated the logic that determines the version string.
We use CMake's PROJECT_VERSION and append the name of the release tag to it in parentheses. So the version is something like "v2.1.0 (2018.03)". If there's no release tag we add the hash of HEAD, eg "v2.1.0 (git-eae4c050cde98fc386dfef0586b43738ceba2d60)". If you don't have git installed then it's "v2.1.0".