tests: Remove incorrect IF
When there is a file called "something.sb" already in the scripts
directory, the install function will only overwrite it, if the tests is
newer. However, this assumption does not hold.

Especially when using the compilation cache and
public/arithmetic_test_scalar and shared3p/arithmetic_test_scalar are
used.